Warn the user if they've encountered an issue with mobile safari.
In Mobile Safari, in iOS 14.5.1 and later, the number of asyncified
calls that can be on the stack at once has been reduced to 100. This
isn't enough for Ren'Py to start, as each Python call takes about
2 stack calls.

There doesn't appear to be a way to work around this regression,
so warn the user and ask them to report it to Apple.